+++ examples/hello_world_experiment.py
@@ -104,4 +104,3 @@ kfp_client.create_run_from_pipeline_func(
 #    namespace="kubeflow-user-example-com",  # your user namespace
 #    experiment_name="my-first-experiment",
 #)
-